Release v1.0 (#91)
* e learning up to date

* e learning up to date

* e learning up to date

* e learning up to date

* build facade to avoid circular dependencies in service

* Course feature (#19) FE addCourse, addClass

* course_feature

* change IClassStudentList

* course_feature

* delete repository

* build frontend

* course feature up to date

* e learning up to date

* Course feature (#7)

* course_feature

* change IClassStudentList

* course_feature

* delete repository

---------


* Dev (#11)

* Course feature (#7)

* course_feature

* change IClassStudentList

* course_feature

* delete repository

---------


* build frontend

* course feature up to date

* e learning up to date

---------


* e learning up to date

* e learning up to date

* e learning up to date

* e learning up to date

* course_feature

* course_feature

* delete repository

* create course

* create course

* update

* update

* update adminCourseController

* update

* course_feature

* change IClassStudentList

* course_feature

* delete repository

* Dev (#11)

* Course feature (#7)

* course_feature

* change IClassStudentList

* course_feature

* delete repository

---------


* build frontend

* course feature up to date

* e learning up to date

---------


* course_feature

* course_feature

* delete repository

* create course

* create course

* update

* update

* update adminCourseController

* update

---------


* update dev

* update courseFacade

* hotfix-null-exception

* Student Service view

* file upload feature

* update dev

* update dev

* update

* my course feature (#22)

* update dev

* update dev

* Student fix

* fix service

* update ClassService

* update dev 6/4/2024

* update form data

* Info Doc of Class

* hotfix update function in firebase

* 404 Page, 403 Page, Loginform.

* update score

* Student register fix

* update

* update score

* update score

* fix infoClass

* update 10 04 2024

* add dao, service of SemesterModel

* hotfix clas service

* update

* update score/ learning

* update

* a big change in service architecture

* update

* update

* Update view_course, my_course

* Student register fix, Teacher register, Tearcher Service View

* update

* Home page; fix 403, 404; doi duong dan css

* a big change in service architecture

* update dev 12 04 2024

* resolve conflict 12 4 2024

* Update course, my_course, download file

* Update

* semester

* paging 50%

* update

* Sua dich vu dang ki mon hoc, lop hoc

* paging 55%

* Trang chu, about us, footer

* dev update

* Sua dich vu dang ki mon hoc, lop hoc - new version

* Update: course filter

* Update

* danh sach sv

* update danh sach sv

* paging for FE

* paging done

* update score

* update dev

* update score

* update score

* fix notFound SV in class, chặn xem ds của sv

* fix errror

* paging for all table avoid table in update site

* update SecondaryId for DAO

* Footer(bao gom About us, contact), trang chu, Rare_fault.

* latest_change_registration

* Chinh sua footer

* udatesacor

* udatesacor

* udatesacor

* fix errror

* update for pagination in DFD

* add semester

* thông báo khi chưa có sinh viên

* thông báo khi chưa có sinh viên

* update dev after meeting 16/04/2024

* config security for admin

* Update nav of (course, list student, input point) , delete courseFacade, update folder image, fix admin-home redirect, update editMode of teacher

* udatesacorggg

* class table, delete class having teacher

* Quản lý đăng ký môn học ở admin

* update

* update

* update search, delete class

* update 18/4/2024

* Update Point, upload excel

* udatesacorooooo

* Update paging, search,...

* Xoá phần chú ý, không hiển thị điểm 15

* Sửa tiêu đề

* update dev

* Feature with file excel (#85)

* init for upload excel

* upload excel and convert to object

* register with Excel file

* register with Excel file 90%

* register with Excel file complete

* register with Excel file complete

* Phân trang danh sách lớp

* excelpoint (#86)

* Dịch lại sang tiếng Việt + disable chức năng ở đăng ký khi không có quyền (#89)

* vietsub (#90)

---------

+++ src/main/java/com/hcmutap/elearning/service/impl/PointService.java
@@ -1,17 +1,25 @@
 package com.hcmutap.elearning.service.impl;
 
-
+import com.hcmutap.elearning.dao.firebase.Options;
 import com.hcmutap.elearning.dao.impl.CourseDAO;
 import com.hcmutap.elearning.dao.impl.PointDAO;
+import com.hcmutap.elearning.exception.NotFoundException;
+import com.hcmutap.elearning.exception.NotFoundInDB;
+import com.hcmutap.elearning.model.ClassModel;
 import com.hcmutap.elearning.model.CourseModel;
 import com.hcmutap.elearning.model.PointModel;
 import com.hcmutap.elearning.service.IPointService;
 import jakarta.annotation.Resource;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageImpl;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
+import java.util.stream.Collectors;
 
-import java.util.ArrayList;
+
 import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
 
 @Service
 public class PointService implements IPointService {
@@ -24,26 +32,65 @@ public class PointService implements IPointService {
     public List<PointModel> findAll() {
         return pointDAO.findAll();
     }
-    @Override
-    public PointModel findById(String id) { return pointDAO.findById(id); }
 
     @Override
-    public void save(PointModel pointModel) {
-        pointDAO.save(pointModel);
+    public List<PointModel> findBy(String key, String value) throws NotFoundException {
+        List<PointModel> pointModelList = pointDAO.findBy(key, value, Options.OptionBuilder.Builder().setEqual().build());
+        if (pointModelList.isEmpty()) {
+            throw new NotFoundException("Not found point with " + key + " = " + value);
+        }
+        else {
+            return pointModelList;
+        }
     }
 
     @Override
-    public void update(PointModel pointModel) {
-        pointDAO.update(pointModel);
+    public PointModel findById(String id) throws NotFoundException {
+        List<PointModel> pointModelList = pointDAO.findAll();
+        if (pointModelList.isEmpty()) {
+            throw new NotFoundException("Not found point with id: " + id);
+        }
+        else {
+			try {
+				return pointDAO.findById(id);
+			} catch (NotFoundInDB e) {
+				throw new NotFoundException(e.getMessage());
+			}
+		}
+	}
+
+    @Override
+    public String save(PointModel pointModel) {
+        return pointDAO.save(pointModel);
     }
 
     @Override
-    public void delete(String id) {
-        pointDAO.delete(id);
+    public void update(PointModel pointModel) throws NotFoundException {
+		try {
+			pointDAO.findById(pointModel.getFirebaseId());
+            pointDAO.update(pointModel);
+		} catch (NotFoundInDB e) {
+			throw new NotFoundException(e.getMessage());
+		}
     }
 
     @Override
-    public  PointModel getPoint(String studentId, String courseId){
+    public void delete(List<String> ids) throws NotFoundException {
+        for (String id : ids) {
+            try {
+                Optional<PointModel> optionalPointModel = Optional.ofNullable(pointDAO.findById(id));
+                if (optionalPointModel.isEmpty()) {
+                    throw new NotFoundException("Not found point with id: " + id);
+                }
+                pointDAO.delete(id);
+            } catch (NotFoundInDB e) {
+                throw new NotFoundException(e.getMessage());
+            }
+        }
+    }
+
+    @Override
+    public  PointModel getPoint(String studentId, String courseId) {
         List<PointModel> pointModelList = pointDAO.findAll();
         for(PointModel pointModel : pointModelList) {
             if (Objects.equals(pointModel.getStudentId(), studentId) && Objects.equals(pointModel.getCourseId(), courseId))
@@ -53,58 +100,100 @@ public class PointService implements IPointService {
     }
 
     @Override
-    public  double getAveragePoint(String studentId, String courseId){
-        PointModel pointModel = getPoint(studentId,courseId);
-        CourseModel courseModel = courseDAO.findById(pointModel.getCourseId());
-        double averagePoint = pointModel.getPointBT()*courseModel.getPercentBT()
-                            + pointModel.getPointBTL()*courseModel.getPercentBTL()
-                            + pointModel.getPointGK()*courseModel.getPercentGK()
-                            + pointModel.getPointCK()*courseModel.getPercentCK();
-        return  averagePoint;
+    public double getAveragePoint(String studentId, String courseId) {
+        try {
+            Optional<PointModel> optionalPointModel = Optional.ofNullable(getPoint(studentId, courseId));
+            if (optionalPointModel.isEmpty()) {
+                return 0;
+            }
+            PointModel pointModel = optionalPointModel.get();
+            Optional<CourseModel> optionalCourseModel = Optional.ofNullable(courseDAO.findById(pointModel.getCourseId()));
+            if (optionalCourseModel.isEmpty()) {
+                return 0;
+            }
+            CourseModel courseModel = optionalCourseModel.get();
+            double averagePoint = pointModel.getPointBT() * courseModel.getPercentBT()
+                    + pointModel.getPointBTL() * courseModel.getPercentBTL()
+                    + pointModel.getPointGK() * courseModel.getPercentGK()
+                    + pointModel.getPointCK() * courseModel.getPercentCK();
+            return (averagePoint / 100);
+        } catch (NotFoundInDB e) {
+            throw new RuntimeException(e);
+        }
     }
 
     @Override
-    public  List<PointModel> getListPointOfStudent(String studentId){
-        List<PointModel> pointModelList = pointDAO.findAll();
-        List<PointModel> pointModelList1= new ArrayList<>();
-        for(PointModel pointModel : pointModelList) {
-            if (Objects.equals(pointModel.getStudentId(), studentId))
-                pointModelList1.add(pointModel);
+    public  List<PointModel> getListPointByStudentId(String studentId) throws NotFoundException{
+        List<PointModel> pointModelList = pointDAO.findBy("studentId", studentId, Options.OptionBuilder.Builder().setEqual().build());
+        if (pointModelList.isEmpty()) {
+            throw new NotFoundException("Not found point of student with id: " + studentId);
         }
-        return  pointModelList1;
+        else {
+            return pointModelList;
+        }
     }
 
     @Override
-    public List<PointModel> getListClassOfStudent(String studentId){
-        List<PointModel> pointModelList = pointDAO.findAll();
-        List<PointModel> pointModelList1= new ArrayList<>();
-        for(PointModel pointModel : pointModelList) {
-            if (Objects.equals(pointModel.getStudentId(), studentId))
-                pointModelList1.add(pointModel);
+    public List<PointModel> getListStudentByClassId(String classId) throws NotFoundException {
+        List<PointModel> pointModelList = pointDAO.findBy("classId", classId, Options.OptionBuilder.Builder().setEqual().build());
+        if (pointModelList.isEmpty()) {
+            throw new NotFoundException("Not found point of student with id: " + classId);
         }
-        return  pointModelList1;
+        else {
+            return pointModelList;
+        }
+//        List<PointModel> pointModelList = pointDAO.findAll();
+//        List<PointModel> pointModelList1= new ArrayList<>();
+//        for(PointModel pointModel : pointModelList) {
+//            if (Objects.equals(pointModel.getClassId(), classId))
+//                pointModelList1.add(pointModel);
+//        }
+//        return  pointModelList1;
     }
 
     @Override
-    public List<PointModel> getListStudentOfClass(String classId){
-        List<PointModel> pointModelList = pointDAO.findAll();
-        List<PointModel> pointModelList1= new ArrayList<>();
-        for(PointModel pointModel : pointModelList) {
-            if (Objects.equals(pointModel.getClassId(), classId))
-                pointModelList1.add(pointModel);
+    public List<PointModel> getListStudentByCourseId(String courseId) throws NotFoundException {
+        List<PointModel> pointModelList = pointDAO.findBy("courseId", courseId, Options.OptionBuilder.Builder().setEqual().build());
+        if (pointModelList.isEmpty()) {
+            throw new NotFoundException("Not found point of student with id: " + courseId);
         }
-        return  pointModelList1;
-    }
+        else {
+            return pointModelList;
+        }
 
+//        List<PointModel> pointModelList = pointDAO.findAll();
+//        List<PointModel> pointModelList1= new ArrayList<>();
+//        for(PointModel pointModel : pointModelList) {
+//            if (Objects.equals(pointModel.getCourseId(), courseId))
+//                pointModelList1.add(pointModel);
+//        }
+//        return  pointModelList1;
+    }
     @Override
-    public List<PointModel> getListStudentOfCourse(String courseId){
-        List<PointModel> pointModelList = pointDAO.findAll();
-        List<PointModel> pointModelList1= new ArrayList<>();
-        for(PointModel pointModel : pointModelList) {
-            if (Objects.equals(pointModel.getCourseId(), courseId))
-                pointModelList1.add(pointModel);
-        }
-        return  pointModelList1;
-    }
+    public Page<PointModel> getPage(String key, String id, int page, int size) {
+        try {
+            List<PointModel> listPoint = getListStudentByClassId(id);
 
+            listPoint = listPoint.stream()
+                    .filter(pointModel -> pointModel.getClassId().equals(id))
+                    .collect(Collectors.toList());
+            if(key!=null) {
+                String lowercaseKey = key.toLowerCase();
+                listPoint = listPoint.stream()
+                        .filter(point -> point.getStudentId().contains(lowercaseKey) ||
+                                point.getStudentName().toLowerCase().contains(lowercaseKey))
+                        .collect(Collectors.toList());
+            }
+
+            long total = listPoint.size();
+
+            int fromIndex = Math.min((page - 1) * size, listPoint.size());
+            int toIndex = Math.min(fromIndex + size, listPoint.size());
+            List<PointModel> pageContent = listPoint.subList(fromIndex, toIndex);
+
+            return new PageImpl<>(pageContent, PageRequest.of(page - 1, size), total);
+        } catch (NotFoundException e) {
+            throw new RuntimeException(e);
+        }
+    }
 }
